More Cleanup

+++ final/users/upload/approve.php
@@ -1,26 +1,39 @@
 <?php
-    require_once 'account/config.php';
+
+    $mysqli=connect();
+
     if($mysqli){
-        $mysqli->autocommit(false);
-        $sql = "";
-        if(isset($_GET['userID']) && isset($_GET['action']) ){
 
-            if($_GET['action'] == "activate"){
-                $sql = "UPDATE `users` SET `status` = 'Active' WHERE `users`.`id` = ?";
-            }elseif ($_GET['action'] == "delete"){
-                $sql = "DELETE FROM `users` WHERE `users`.`id` = ?";
-            }
+        if(isset($_GET['userID']) && isset($_GET['action']) && isset($_GET['status'])){
 
-            if($stmt = $mysqli->prepare($sql)){
-                $stmt->bind_param("i", $_GET['userID']);
+            $sql="";
+            if($_GET['status'] == "Active"){
 
-                if($stmt->execute()){
-                    $mysqli->commit();
+                if($_GET['action']=="accept"){
+                    header("Location: localhost:63342/gentelella-master/final/pages/users/admin/user-details.php");
+                }elseif ($_GET['action']=="decline"){
+                    $sql = "UPDATE `users` SET `status` = 'Suspended' WHERE `users`.`id` = ?";
                 }
+
+            }elseif ($_GET['status'] == "Suspended" || $_GET['status'] == "Pending Confirmation"){
+
+                if($_GET['action']=="accept"){
+                    $sql = "UPDATE `users` SET `status` = 'Active' WHERE `users`.`id` = ?";
+                }elseif ($_GET['action']=="decline") {
+                    $sql = "DELETE FROM `users` WHERE `users`.`id` = ?";
+                }
+
             }
+
+            $stmt=getStatement($mysqli, $sql);
+            $stmt->bind_param("i", $_GET['userID']);
+
+            $mysqli->autocommit(false);
+            executeUpdate($stmt);
+            $mysqli->autocommit(true);
         }
 
-        $mysqli->autocommit(true);
     }
-    header("Location: manage-users.php");
+    header("Location: localhost:63342/gentelella-master/final/pages/users/admin/manage-users.php");
+
 ?>
